libnetwork: sandbox.setupDNS() don't parse host resolv.conf if not needed
The existing code was always parsing the host's resolv.conf to read
the nameservers, searchdomain and options, but those options were
only needed if these options were not configured on the sandbox.

This patch reverses the logic to only parse the resolv.conf if
no options are present in the sandbox configuration.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>