daemon/graphdriver/overlay2: remove deprecated overrideKernelCheck
Commit 955c1f881ac94af19c99f0f7d5635e6a574789f2 (Docker v17.12.0) replaced
detection of support for multiple lowerdirs (as required by overlay2) to not
depend on the kernel version. The `overlay2.override_kernel_check` was still
used to print a warning that older kernel versions may not have full support.

After this, commit e226aea280efdc18f887e853b5eefd763b602937 (Docker v20.10.0,
backported to v19.03.7) removed uses of the `overlay2.override_kernel_check`
option altogether, but we were still parsing it.

This patch changes the `parseOptions()` function to not parse the option,
printing a deprecation warning instead. We should change this to be an error,
but the  `overlay2.override_kernel_check` option was not deprecated in the
documentation, so keeping it around for one more release.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit e35700eb5018ff14d91bf5e46f39e65cfe711b6f)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>