Avoid iptables lock error for endpoint port mapping
On Ubuntu Focal the iptables implementation returns immediately with an
error if the xtables lock is held.  This is a change in behavior that
has exposed a latent issue in Docker's implementation.

After moving to Focal hosts in our CI environment, contention on the
iptables xtable lock is leading to the following error message (these
failures are easy to trigger and frequent):

```
  Error response from daemon: driver failed programming
  external connectivity on endpoint foo
  (4ee313e3bf3f375e70c3ee5d00b9000523cb80e5fb29b9ccff565913c74ab6ec):
  (iptables failed: iptables -t nat -A POSTROUTING -p tcp -s 172.16.128.7
  -d 172.16.128.7 --dport 8081 -j MASQUERADE: Another app is currently
  holding the xtables lock. Perhaps you want to use the -w option?
```

The source of that error message is endpoint.sbJoin in
libnetwork/endpoint.go.  It calls:

```
  endpoint.sbJoin
   driver.ProgramExternalConnectivity (libnetwork/drivers/bridge/bridge.go)
    bridgeNetwork.allocatePorts
     bridgeNetwork.allocatePortsInternal
      bridgeNetwork.allocatePort
       PortMapper.MapRange
        PortMapper.AppendForwardingTableEntry
         PortMapper.Forward
          ChainInfo.Forward
           iptable.ProgramRule
            iptable.Exist
             iptable.exists
              iptable.existsRaw
```

`iptable.raw` calls iptables with `--wait` if supported, but
`iptable.existsRaw` does not.  The invocation of iptables in `existsRaw`
has not changed since 2017, but the OS behavior has.  In Ubuntu 20.04
(Focal) the call to `iptables -S` will fail immediately if the lock is
held by another process, this is not the case in earlier (Bionic) or
later (Jammy) releases.

This patch adds the `--wait` option to this invocation of iptable (or
if that is not supported the `bestEffortLock` is used) when listing
existing rules.

Signed-off-by: Robert C Jennings <rcj4747@gmail.com>
Fixes: #44331