Guard exec with empty command when ip6tables not present
ip6tables is treated as an option executable, unlike the iptables
command, but there are no guards against calling exec.Command with an
empty string for the executable name if an ipv6 rule is provided.  This
patch adds checks where ip6tablesPath is referenced to allow for more
graceful and helpful failure.

Signed-off-by: Robert C Jennings <rcj4747@gmail.com>