Fix swagger docs for ImageSummary
Right now a simple instantiation of a github docker-java DockerClient
hitting the unix socket at /var/run/docker.sock provides an instance of
each of these three values being null. (This is on Docker Desktop 4.9.1
on OSX, engine version 20.10.16.)

They should not be marked as x-nullable: false if they can be null.

Signed-off-by: C.R. Drost <c.r.drost@gmail.com>