windows: daemon: allow pulling/running of newer images under Hyper-V.
Following the addition of Hyper-V isolation, it should be possible for
Windows hosts to run container images with older/newer OS versions given
a Windows host with the RS5 updates and above.

This patch includes the following changes:
- [distribution/pull_v2] enable pulling images newer than the host on RS5+
- [daemon/create] add version checks which prevent running newer/older
  images unless using Hyper-V isolation

Previously, pulling a Windows image with a newer OS version than the
host was prevented by moby#36327, whose behavior is preserved on
pre-RS5 hosts.

Signed-off-by: Nashwan Azhari <nazhari@cloudbasesolutions.com>