vendor: github.com/armon/go-radix v1.0.1-0.20221118154546-54df44f2176c
Previously we had to use a replace rule, as later versions of this
module resulted in a panic. This issue was fixed in:
https://github.com/armon/go-radix/commit/f30034d78803af4786015a5a3a265a40b5837a70

Which means we can remove the replace rule, and update the dependency.
No new release was tagged yet, so sticking to a "commit" for now.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit a2d758acc9a46d6f8a25a5b72950b75f90fe4f06)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>