[22.06] daemon/disk_usage: Use context aware singleflight
This is an adaptation of dec81e489fdc493a4702d035f271814c103a88aa,
adjusted for the 22.06 branch, which does not yet have the context
passed through to daemon.Containers, and does not have the containerd
implementation.

Backporting this code, as the code that it is replacing was newly
introduced in 22.06 through 135cec5d4dd31bb1ce87432c119bb6160d45747c,
and we may not want to be shipping obsoleted code with the release.

-----

The singleflight function was capturing the context.Context of the first
caller that invoked the `singleflight.Do`. This could cause all
concurrent calls to be cancelled when the first request is cancelled.

singleflight calls were also moved from the ImageService to Daemon, to
avoid having to implement this logic in both graphdriver and containerd
based image services.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>
(cherry picked from commit dec81e489fdc493a4702d035f271814c103a88aa)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>