daemon/config: rename CommonTLSOptions to TLSOptions
This type was added in 677a6b3506107468ed8c00331991afd9176fa0b9, and named
"common", because at the time, the "docker" and "dockerd" (daemon) code
were still in the same repository, and shared this type. Renaming it, now
that's no longer the case.

As there are no external consumers of this type, I'm not adding an alias.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>